Created standard query filter, chainable with other filters
